backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
apturl==0.5.2
attrs==25.1.0
awscli==1.22.34
bcrypt==3.2.0
beautifulsoup4==4.13.3
blinker==1.4
boto3==1.35.71
botocore==1.35.71
Brlapi==0.8.3
bs4==0.0.2
certifi==2025.1.31
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
duplicity==0.8.21
ecdsa==0.19.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.11
fasteners==0.14.1
fonttools==4.55.3
future==0.18.2
h11==0.14.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
jmespath==1.0.1
joblib==1.4.2
keyring==23.5.0
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.10.0
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.2.0
oauthlib==3.2.0
olefile==0.46
opencv-python==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
paramiko==2.9.3
passlib==1.7.4
patsy==1.0.1
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.4.8
pycairo==1.20.1
pycocotools==2.0.8
pycups==2.0.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
roman==3.3
rsa==4.8
s3transfer==0.10.4
scikit-learn==1.6.1
scipy==1.15.1
SecretStorage==3.3.1
selenium==4.28.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.46.1
statsmodels==0.14.4
systemd-python==234
threadpoolctl==3.5.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
undetected-chromedriver==3.5.5
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.34.0
wadllib==1.3.6
websocket-client==1.8.0
websockets==14.2
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
